Fix possible buffer overflow in getFieldAddress
When StreamProtocolParser::Protocol::compileFormat finds a
%(other_pv)d format specifier it calls StreamEpics::getFieldAddress.
getFieldAddress will first try to interpret other_pv as a field name
checking for the existence of this_pv.other_pv. If it doesn't find
such PV it will then assume the format points to other_pv.VAL.

If both this_pv and other_pv have long names they will cause an
overflow on the buffer 'fullname', which has a size of PVNAME_SZ+1 = 61.

This fix replaces the sprintf call with snprintf.